refactor: deprecated API migration[#57]
